Add files via upload
Logical Categories go over data and the logic for data wrangling. 
Finalproject.py does a multivariable logistic regression of the data
project gradient.py does a global gradient descent optimization of the logistics regression. 
adultdata.csv, is the cleaned data. UIC machine learning repository is the source of the data.